Add `SmartSemicolonInsertion` auto edit
This auto edit always moves the semicolon to the end of the line unless
the cursor is inside a for comprehension - the only place in Scala where
the usage of semicolons is a syntax feature.

The "auto insert braces at corret position" feature of the Java editor
is no longer supported by the Scala editor because there are very few
cases where this feature really makes sense. Most of the time one can
use a template to generate code where the opening brace is at the
correct position.

Fixes #1002056