Fix Luna build by using the AJDT dev version of the weaving hook.
AspecJ has a new release in their /dev update site, with a different
version of the weaving hook. We upgrade to this one to fix our builds.

This is (probably) brought in as a transitive dependency from aspectj.runtime,
and I couldn’t find a way to use the weaving.hook available in the release
Luna repository (there is one there as well).

We might investigate whether we should use the AspectJ release from the
Spring Tool Suite (since that’s the most likely conflict we’ll ever have),
at least until there’s a stable release of AJDT for Luna (none yet).