Restore correct cursor position when cursor is before removed range
In case there exists a multi text edit, which removes text before and
after the text selection, the cursor wasn't correctly restored.

The implementation need to find all regions that exist before the text
selection and only subtract them from their size from the old cursor
position in order to create the new position. Because the implementation
also needs to find out if the text selection intersects with a changed
region it is easier to find all regions that exist after the selection
and subtract their size from the entire size of the change.