Fixes problem build disabled after clean
Following the merge of #767, doing a clean with 'Build Automatically' on, would
not generate any binaries.
The reason was the job scheduled to update the markers after the classpath check (normal after
a clean) was interrupting the build job. The fix is to put both job at the same priority level,
so they don't interrupt each other.
Also fixes the fact that EclipseBuildManager#clean was called twice for each project.