Fix AIOOBE in ScalaWordFinder
The exception could occur because an arbitrary Seq can be passed to the
method findWord and not just some special wrappers that are defined in
the overloaded alternatives. Because there were also no offset checks,
the result would have been completely useless if this method would have
succeeded without an exception.

Therefore, this commit introduces offset checks and returns None if the
offset is invalid. In order to return an Option, the API of findWord
had to change. Because the word finder is used in some AspectJ code, a
new method `getWord` was introduced in a former commit, which returns
null instead of a `None`.

Another change is that findWord doesn't accept a Seq anymore, but an
IndexedSeq. Otherwise a passed linked list would result in an
inefficient implementation.

Fixes #1002238