Remove `HoverAspect` and `IScalaWordFinder`
I'm not entirely sure if the removal of `HoverAspect` doesn't break
anything. It is definitely not used in Scala hovers, because the Scala
IDE has its own implementations. It may be used by some Java hovers, but
I couldn't find one that actually uses it. In the worst case this change
means that some Scala elements are no longer shown in Java hovers.