Remove status handler for missing Scala library, and instead use a marker resolver.
This removes the pop-up dialog for a missing Scala library when trying to instantiate
the presentation compiler. Instead, the error marker in the Problems View, which was
anyway created, now has a marker resolver (enabling QuickFix — Ctrl-1), just like
Scala Installations errors.

This makes it possible to remove the IScalaProject dependency from PresentationCompilerProxy,
which should be able to manage any presentation compiler (with or without an associated project).