Display instantiated type in signatures, for hovers
The type displayed in signatures for completions is still the uninstantiated type, i.e.

l = List(1,2,3)

| Hover              | Completion       |
|--------------------+------------------|
| l.head : Int       | l.head : A       |
| l.tail : List[Int] | l.tail : List[A] |
|--------------------+------------------|

The intent is to make clear that the 'tail' member (on which the focus
is in completion) is *defined* on generic lists, whereas in a hover,
the focus is on the type of a given expression, which is therefore
instantiated (in the case of tail, a List[Int]).