Remove the hard-coded InteractiveCompilationUnit from `ScalaHover`
The associated compilation unit may change (see #1002251) when the editor
is reused. On top of that, the structured editor in the Play plugin needs
a no-arg constructor (it’s instantiated through an extension point).

This change allows the Play plugin to use plain Scala hovers, including
all the goodies (configured font and CSS, scaladoc, etc).