Add ScalaEditorAnnotation
This annotation helps to to simplify the new quickAssists extension
point. Annotations are now directly passed to the quick assists. This
means that they no longer have to search for them by themselves, which
speeds ups the whole process.

Another advantage of this change is that no dependency to JDT for the
quickAssists extension point is used anymore.