Adapt quick assist test suite for new quickAssists extension point
Most of the tests are bad and because they rely on UI behavior they
can't even be run by the build server. But the tests exist and I don't
want to remove them. If someone has more time available it would be
nice if this person could move tests completely away from UI classes.