Faster project loading in classpath computation
Loading projects is recursive and in case where:

A -> (depends on) B, A -> C, B -> D, C -> D

D is loaded twice.

In some big projects this causes a lot of wasted time when projects are
loaded dozens of times.

This commit introduces cache for already computed projects thet prevents
multiple loading of single projects.