Handle src/bin folders that are the same as project root folders
It is possible to create source and binary folders that are the same as
the project root folder. The easiest way to do this is to use the "Use
project folder as root folder for sources and class files" option in the
new Scala project wizard.

Scala IDE is unable so far to handle these folders because the Eclipse
API itself seems to be unable to handle these folders. The common
operation to retrieve an IFolder is

    project.getFolder(path)

which throws an exception in case path is the same as the project path.

This commit fixes the problem by not calling the `getFolder` method but
by creating an `IFolder` manually or by reusing the project object
directly.

Fixes #1002332, #1002146