Mark a recently flaky test as, erm, Flaky.
This one failed three times in PR validation this week. The CI
farm has been pretty crowded with a swathe of scala-js builds,
so that might have flushed out a race condition in this feature.

For now, I have placed the test under the watchful and patient
eye of the FlakyTest retry mechanism.