Handle src/bin folders that are the same as project root folders
It is possible to create source and binary folders that are the same as
the project root folder. The easiest way to do this is to use the "Use
project folder as root folder for sources and class files" option in the
new Scala project wizard.

Scala IDE is unable so far to handle these folders because the Eclipse
API itself seems to be unable to handle these folders. The common
operation to retrieve an IFolder is

project.getFolder(path)

which throws an exception in case `path` is the same as the path of
`project`.

The problem is fixed by not calling `getFolder` inside of `ScalaProject`
but by reusing an existing reference to the `IContainer` that already
references the correct location on the filesystem.

Fixes #1002146
(cherry picked from commit 6f434e42fd8554eb99b884276d45d2b1f5d26948)