Various improvements for "Copy Qualified Name"
Switch to CompilerSupportTests so that we can get rid of separate
source files for the unit tests.

Fix a race condition when creating compilation units for unit tests.

Add additional test cases for method arguments (they are currently
skipped, as the implementation is still missing).

Access Symbols in the PC thread.

Organize imports.