Fix regex to match \r\n before \r or \n.
The template context was previously splitting templates by the regex \r|\n|\r\n. Since | is eager, not greedy in Java, when applied to strings containing '\r\n' it would produce two matches, which would then be re-joined with two line endings. This change sorts longer line endings to the start of the regex to ensure that "\r\n" is matched as a single unit.

Fix #1002303.