Treat empty groups separately in auto edit templates
It can happen that an user wants to create a template that contains
multiple empty markers (${}). An example of such a template is:

    def ${name}(${}) ${} = {
      ${cursor}
    }

In this case the idea is to be able to update each region separately
when test is inserted into the editor, i.e. only the empty marker that
is currently edited should be updated. However, the template API doesn't
allow us to disambiguate between these empty markers, it treats them as
one group. Therefore we have to manually untangle them and pack each
empty marker into its own group.