Do not cycle in linked mode model when parent exists
A parent exists when a template is expanded inside of an already
existing expanded region. For example, if an user inserts multiple
opening parentheses the result may be (((^))) where ^ denotes the
cursor position. When the user now presses tab, it is very likely that
they want to jump to the next closing parenthesis which is only possible
if cycling is disabled when parents exist.