T1002387 Remove an option to close SPCs when editors are open
There are known cases when this (by default disabled) option leads to
memory leaks. Since the way in which dependencies of the presentation
compiler are freed hasn't been corrected so far, it's better to remove
this option. It can be brought back when the related problems will be
solved and someone will still consider it as something helpful.

This commit removes the preference stored previously in the preference
store, the checkox from GUI and underlying logic. It also updates
tests.