Macro compile first feature.
Problem: originally there was a problem with compiling code when macro
is defined. Then macro code had to be moved to separate project. On other
hand building process of dependent projects is interrupted if a
compilation of any project fails. It is undesirable behavior when for
example compiling error is met during test compilation.
To fix it a functionality of the separate scopes compilation has been
implemented.

Compilation scope is a set of sources which are compiled together and a
product of compilation is stored in common output folder.

There are defined (as a convension) 3 scopes:
- macros with default source path: <project root>/src/macros* and
  especially for Play projects <project root>/conf*
- tests with default source path: <project root>/src/test* and (because of
  Play projects) <project root>/test*
- main which gets all source paths which don't match macros and tests scopes
(* is a wildcard so for example to tests scope are assigned sources from
folders src/test, src/test-my-special)

The scopes are compiled in order: macros, main, tests. The compilation is
conditional so if any scope compilation fails then further scopes are not
compiled. So when macros compilation fails then neither main nor tests is
compiles. When macros compiles successfully and main fails then tests scope is not compiled.

When given eclipse project (let name it B) depends on other project (let
call it A) then there are introduced additional dependencies between
projects' scopes:
- macros of B depends on macros and main of A. So macros of B requires the
  correct compilations of macros and main scopes in project A
- main of B depends on macros and main of A
- tests of B depends on macros, main and tests of A

The feature is controlled by 'useScopesCompiler' exposed on 'build
manager'tab in Scala->Compiler preferencies menu.

Remark: commit contains all reworks and refactorings done along the implementation

Fix #1001633