Basic support for Hot Code Replacement in debugged JVM
This commit adds the possibility to modify and compile code, when an
application is run in a debug mode, and to have these changes visible
and taken into account by the debugged VM without restarting the
application. This feature can be turned on/off (right now it's turned
off by default). User can decide whether files containing compilation
errors should be skipped. Stack frames affected by changes can
be automatically dropped. Sometimes users don't like automatic
dropping frames after HCR and they prefer to do that manually (see HCR
in IntelliJ). Right now there's added also such a possibility. On the
other hand it's a question whether it should be available, if we'll
know it can cause problems. Therefore, it may be removed in the future,
if the HCR with automatic dropping frames would be much more reliable.
There are also added several integration JUnit tests.

Note that HCR is limited by JVM's support for replacing classes. Its
goal is to allow to experiment with the code inside methods/blocks
rather than to develop the whole application in the interactive mode.
Changes of methods' signatures etc. don't work - similarly to other
HCR implementations from Eclipse or other IDEs.

The implementation is inspired by the HCR support for Java. Scala
debug target, threads etc. don't extend the default Java/JDI model and
build their own one. Therefore, they are skipped when processing
events by various Java-focused managers.

There's added ScalaHotCodeReplaceManager which implements HCR support
using the Scala model. Every ScalaDebugTarget has its instance of
manager which is also the listener waiting for events related to
changed resources. When we get an event about changed class files, we
get byte arrays containing new versions of these files and replace the
old ones in VM. After this operation it's needed to rebind frames
(optionally also drop the obsolete ones) and to refresh breakpoints.

Each from several added tests contains a fragment where the code is
modified and recompiled. To ensure that all events and messages are
correctly propagated after the additional incremental build, it was
needed to use Thread.sleep with 500 millis. I know it's quite ugly
solution but it seems to work well and there are not many such tests.

There are serious reasons why this implementation is called the basic
support:

- Automatic dropping frames is based on the information which frames
are marked as obsolete.

- We get this information after performing HCR but it can cause
problems. According to the comment from Eclipse's HCR implementation
for Java, JVM is faulty when there are frames rendered as obsolete.
Therefore, in Java implementation they try to find affected frames
and drop them before performing HCR. We should do the same thing
because VM sometimes crashes after HCR and some additional operations.
This is the serious blocker. Anyway even current implementation can be
quite helpful in many situations. Except JUnit tests it was used to
debug e.g.some multithreaded application using futures.

- There's also another problem. After performing HCR sometimes
already existing breakpoints behave strange. I had situations where
from time to time I was getting a JDI event for the line breakpoint
with the different location than the line specified for this
breakpoint (sometimes it was one line before or after the expected
one). I was investigating this for a while but didn't figure out what
to do with that. Also I wasn't able to reproduce such a behaviour in
tests. It's something, what should be checked once again after the
addition of the new approach to drop affected frames. In case of
problems it's always possible to recreate given frame manually, drop
to the above frame and go to the expected breakpoint once again.
Note: it's sometimes hard to hit breakpoints in Scala IDE even without
HCR (they quite often are nondeterministic from the end-user's point
of view of). Hence, it's not so simple to test this. It's not possible
to fix everything, which basically is not related to this particular
feature, but is also used when debugging.

- Dialogs shown when replacing classes is not supported by VM or fails
don't contain buttons to terminate or restart debug session. Currently
there's only an adequate message.