Don't spawn new mark occurrence job when one is already running
The best solution would be to cancel an already running job before
spawning a new one. Unfortunately it is not possible to cancel an
already started invocation of the presentation compiler - only the job
itself can be canceled, which shouldn't have any effect because it
isn't doing a lot of work.

Therefore the current solution is to not start a new job until another
one has been completed. The downside is that the mark occurrence feature
is not invoked when it could be invoked, namely when a new identifier is
selected but the mark occurrences job for the previously selected
identifier has not yet been completed.

Overall I think this would only happen if mark occurrences is really
slow but in this case the user doesn't get their highlighting anyway on
time, therefore for them the feature is still broken. But at least there
is no chance anymore to start a lot of jobs, which will not complete on
time.

If mark occurrences is fast, it should already be completed before users
decide to select a different identifier and therefore this change
shouldn't affect them.

Fixes #1002417