Refresh variables view after evaluation
An expression evaluated in expression evaluator can contain the
assignment to var or call a method having side effects like
changes in some internal or global state. That's why we need to
refresh the variables view - to have it up to date after such possible
modifications.
Unfortunately there's one related inconvenience. If user had some
non-top-level node selected in the variables view, he can loose the
current selection. Then it's needed to select this element once again.
Therefore, there's also added an option in preferences which allows to
turn this refreshing off. It's a tradeoff.