Smarter HCR-related tests
So far there was Thread.sleep used with some hardcoded value to ensure
that all events were propagated correctly after the build. In general
it was:

a) buld project with changes
b) there are some post-build actions performed and events are sent in
a background so wait 500 millis

It was causing problems in case of Scala's Jenkins for which 500
millis was too low value. Also it's not the best idea to wait 500
millis (always) even if it's not needed.

Therefore, this commit provides a bit smarter solution. From now on
HCR-related publisher publishes the HCRSucceeded event, when HCR is
finished. In tests we take into account what happens under the hood
in aforementioned point b. It is:

1. perform HCR
2. send a message to the thread-related actor => some asynchronous
processing (at the end we fire an event)
3. send a message to the breakpoints-related actor => some
asynchronous processing
4. notify subscribers about finished HCR (asynchronously)

In tests we wait for the events from points 2 and 4 (with some quite
big time limit, what should be helpful in case of Jenkins), and after
that check whether current stack frame is correct, variables have
proper values and so on. Unfortunately there's no event etc. from
BreakpointSupport so we check nothing related to the third point.

From now on the test checking whether disabling HCR works correctly
checks only that ScalaHotCodeReplaceManager is not created for
ScalaDebugTarget. It's enough.

There's also added to ScalaHotCodeReplaceManager the detailed
description how HCR works in case of errors in Java or Scala code
(the differences are a result of differences in deltas which we get
in events).