Fixes StackOverflowException in ScalaIndenter
The main problem is an exponential blowup of a regex. Another problem is
that the val `restOfTheLine` doesn't really contain the rest of the line,
but the start of the string literal until the offset, what in this case
is the end of the string literal. Another problem is that the regex is
too complicated.

The easiest way to fix the problem was to simplify the regex and to stop
the regex computation as soon as a newline character occurs, which
happens because `.` doesn't match newline characters.

No tests included because a test suite doesn't exist. The following
should be the behavior (^ is the cursor before/after `<enter>` is pressed):

    def f = ""^
    //==>
    def f = ""
    ^

and

    def f = ^""
    //==>
    def f =
      ^""

Fixes #1002433