Accept escaped signs and unicode literals in chat literal detection
Regex for unicode literals is copied from `StringTokenScanner`. Invalid
escaped sign are also accepted, as valid char literals because it seems
to be more likely that no indentation should happen after such compile
errors.

Valid/Invalid escape sign:

    def u = '\x'^
    // ==>
    def u = '\x'
    ^

Valid unicode literals

    def u = '\u0fff'^
    // ==>
    def u = '\u0fff'
    ^

Invalid unicode literals

    def u = '\u0ff'^
    // ==>
    def u = '\u0ff'
      ^

It is questionable if the last case is correct or not, but it is such an
unimportant case that it should not really matter in practice.