Enable compiler checks at build time
`-Xfatal-warnings` has been enabled in order to guarantee that no new
code smells are introduced into the codebase.

`-deprecation` has been disabled because there are were some deprecation
warnings reported that couldn't easily be fixed. They are all related to
JDT/Eclipse interoperability.

All `-language` flags has been enabled because we use them anyway and
importing them everywhere is annoying.

`-Xlint:missing-interpolator` has been disabled, because it produces
some false positives in our codebase. Actually I'm not sure, if this a
problem in scalac or in the scala-maven-plugin because in when calling
scalac through the command line with a little example all the false
positives are gone. It is a shame to disable this warning because it is
in general a very useful one.