Add Unit return type to procedures
All changes are done automatically with the following command:

    find . -iname "*.scala" -print0 | xargs -0 perl -i -pe 'BEGIN{undef $/;} s/(def (?:(?!this)\w+)(?:\([^)]*\))*)( *{)/\1: Unit =\2/gsm'

I don't know perl well, therefore I don't know why the `BEGIN` block is
needed. Without it a multi line def is not recognized by the regex. It
is explained here:

    http://stackoverflow.com/a/1030819/2621536

After the above command, there were still some defs with procedure
syntax in the codebase, namely abstract defs. In order to get rid of
them the command had to be changed a little bit:

    find . -iname "*.scala" -print0 | xargs -0 perl -i -pe 'BEGIN{undef $/;} s/(def (?:(?!this)\w+)(?:\([^)]*\))*)$/\1: Unit/gsm'

The difference here is that after detecting the parameter list, we
immediately check that the line ends.

However, even after the second command, there were a few occurrences of
procedure syntax left. These were defs that take a function or a tuple
as an argument and therefore the couldn't be detected by the regex.
Nevertheless, fixing this manually was trivial.