Consider typechecking phase for timeout of save actions
So far only the running time of the save action itself was considered in
the timeout. With this change, the typechecking phase of the compiler
that is needed for every compiler driven save action is also considered
in this timeout.

Another change is that after a timeout occurred, all further compiler
driven save actions are aborted. This should be the better behavior
because it is not likely that the time that is needed to typecheck and
refactor a file largely differs between different save actions. Therefore
it is unrealistic that further save actions can succeed. Also, a save
action can not be aborted (because interrupting the compiler thread is
not supported) and therefore another call to the compiler first have to
wait on the previous request to be completed, which makes it even more
unrealistic that further save actions can be completed successfully.

Fixes #1002308