Let auto edits handle changes where text is already selected
In cases a user has selected some text and then presses a key the wrong
behavior happened. So far the selected text was replaced with the
pressed key. Instead we not only want to replace the selected text but
also call the auto edit that belongs to a key.

Example: If the user inserts an open parenthesis, the closing
parenthesis should also be added after the selected text is replaced.