Move save actions and auto edits to their own object
They need to be available very early in initialization of Scala IDE (due
to the fact that they are accessed in preference initializers). By
moving them to their own object that does not have further dependencies
we ensure that we don't get into any initialization issues. Previously
the logger was not available because its initialization was not yet
completed at the point where save actions or auto edits were
initialized.