Actors don't work well with JDI locks
The easiest way to have common JDI locking is just to use a lock
on JDI access. It makes some troubles in actors because actors'
threads can be terminated when a lock is entered and then the lock
is hold.
So the lock section in actors code is moved to Future which uses
system thread pool with a cleaning mechanism.