JDI synchronization excluding actors
The easiest way to have common JDI locking is just to use a lock
on JDI access. It makes some troubles in actors because actors'
threads can be terminated when a lock is entered and then the lock
is hold.
But because debug actors are moved to Futures the JDI sync in actors
is skipped.