Handle bundle locations correctly
`getLocation` shouldn't be used, it is not always correct. Furthermore,
it is cumbersome to work with it. `FileLocator.getBundleFile` is used
instead, it returns a `java.io.File` that points to an existing file.

Furthermore, more logging is enabled. This makes it much easier in
future if any errors occur.