Resolve execution conflict between surround selection and close pair
The problem was that when both surround selection and the close pair
auto edit were enabled, one of them is preferred over the other. To
solve this conflict some priority had to be introduced. Right now an
easy was is chosen - the close pair auto edits ensure that the surround
selection auto edits are not enabled.

Furthermore, `SurroundSelectionWithAngleBrackets` had to be introduced
to allow the corresponding close pair auto edit to check if it exists.

Fixes #1002488