Remove stale HierarchyResolver.remember around advice.
The around code currently breaks running JUnit tests in Java projects that use
lambdas. This is because the cast to SourceTypeBinding fails on lambda bindings.
Eclipse suppresses the ClassCastException and tests just do not run.

As far as I can tell, there is actually no Scala IDE-specific code in this
"remember" advice anyway. It was added in 2009 by Miles, and looks exactly like
the code it's replacing in older releases of Eclipse:

https://github.com/eclipse/eclipse.jdt.core/blob/R3_2_maintenance/org.eclipse.jdt.core/model/org/eclipse/jdt/internal/core/hierarchy/HierarchyResolver.java#L347

(E.g. unlike most advice, it was not calling proceed/etc., but instead was just
a raw copy/paste of the original code. Probably well intentioned while trying
to get things to work.)

However, newer versions of Eclipse have added an instanceof check to handle lambads:

https://github.com/eclipse/eclipse.jdt.core/blob/master/org.eclipse.jdt.core/model/org/eclipse/jdt/internal/core/hierarchy/HierarchyResolver.java#L413

Which means this copy/paste of the code is now stale.

Instead of updating this code to match newer version of Eclipse, given
it looks unneeded, I've just removed it.