Simplify logic that modifies Scala installation
Previously the logic was too complicated and also led to a bug that made
it impossible to switch from one compatibility mode to another one (for
example configure 2.10 support instead of an already configured 2.12
support).

The logic is simplified insofar that the Scala installation is changed
before `InstallationManagement.getCompatibilityMode` is called at all.
This way, we ensure that we only check for the desired Scala
installation and not the one that is already configured (and therefore
should be changed).