Skip expensive buildStructure calls when `NO_AST` and no `resolveBindings` are needed.
This code path is taken by `Save`, on the UI thread, and
slow build structure can block the UI for several seconds while saving.

In order to not lose functionality, the structure is now built after each reconciliation. It should be fast (the whole file was already type-checked), and on a background thread.

Fixed #1002412