Scala debugger configuration
Because of resource consumption by Async Debugger Tracer
(it stores in memory the history of stack traces gathered
across executions of deferred code blocks like actors both
akka and scala and scala Futures too) there is a need to
allow a used to enable or disable this feature.
So the Scala Debugger tab has been added to 'Launch Configuration'
to control if async debugger garners or not async stack traces.
It has been added to following Run/Debug Launch Configurations:
- Eclipse Application with Equinox Weaving
- JUnit
- JUnit Plug-in Test
- Remote Java Application
- Scala Application
- ScalaTest