Bang operator not working
There are two things fixed in this commit:
- command ids didn't reflect the id in command definition
- command must be associated with debug context
To explain the second it is enough to look at `StepMessageOut` class.
The first element of selection must be `ScalaStackFrame` and continue
action should run debugging further. In buggy solution the first element
of selection is `AsyncStackFrame` and action does not continue debugging.