Add update-site for scala-refactoring
Right now, the scala-refactoring repo contains an update-site, which is
used by the Scala IDE build. Since the update-site just clutters up the
scala-refactoring repo, the idea was to move the update-site to the
scala-ide repo. This way the scala-refactoring repo is easier to
understand. The scala-ide repo gets more complicated but we can hide all
of these artificial update sites in a sub directory instead of keeping
them in the root directory as it is done right now.

Another change that was required to make this up and running can be
found in the nightlies of scala-refactoring. So far a P2 update site was
published by our build server. This has changed in so far that a nightly
of the scala-refactoring library is published, which can also be
consumed by other people.