Replace usage of scala-pickling with json-io
The latter is a small Java library, which inspects objects at runtime in
order to (de-)serialize them. This may be slower than the approach taken
by scala-pickling but it requires less code and it is faster to compile.
Not to mention that the code is less susceptible to compiler errors.