Add Import from current file switch
A new flag on Expression Evaluator preference page has been added
to toggle import of packages. This feature is introduced to skip
the paste of import because Toolbox is not able to handle a case
when package is nested in package (see definition of
NestedPackagesSample.scala packaging definition at top of file).
Test `shouldFailBecauseNestedPackagesNotSupportedByToolbox()` should
pass when Toolbox is fixed.