Improve handling for corrupted statistics file
Previously, we just logged the error in case an error happened. Now, we
log an understandable error message and also log the corrupted JSON
content. The file is then being overwritten. It would be better to fix
the data corruption but we can't do that automatically and we don't want
to bother users to do it by themselves.

Users that care about the data a lot and want to fix it manually, still
can get the corrupted data out of the log and merge it back to the
statistics file. Not very nice to do for anyone but I'm too lazy to
write a JSON merger now. I think about it again when someone complains
about it.

Fixes #1002600