Simplify trait CompilerSupport
Previously this trait contained lots of definitions that were copied
from scala-refactoring. By reusing the functionality in
scala-refactoring, the logic could be simplified a lot. The only
downside is that scala-refactoring creates text change objects of the
wrong class and we have to do a conversion to the expected class. But
this shouldn't hurt much, overall there should only be a few objects
that need to be converted this way.