Remove unused imports in core project
These imports were not seen so far, since the compiler doesn't report
them in the most actual nightly build of scalac. However, the newest
version of scalac includes the following fix:

  https://github.com/scala/scala/pull/4899

If the IDE is build on top of a scalac SNAPSHOT that includes this fix,
then it reports the unused imports that are removed by this commit.